fix: endpoints arreglados salvo el delete
