+++ frontend/src/router/index.js
@@ -2,7 +2,6 @@ import Vue from 'vue'
 import Router from 'vue-router'
 import Home from '@/components/Home'
 import AuthPage from '@/components/AuthPage'
-import MisMascotas from '@/components/MisMascotas.vue'
 
 Vue.use(Router)
 
@@ -17,11 +16,6 @@ export default new Router({
       path: '/auth',
       name: 'auth',
       component: AuthPage
-    },
-    {
-      path: '/mis-mascotas',
-      name: 'MisMascotas',
-      component: MisMascotas
     }
   ],
   mode: 'history'
